Update index.d.ts
The nengi example repos use these Instance and Client functions. As far as I can tell, these still need to be available for adding and removing entities when new players join / leave.

https://github.com/timetocode/nengi-2d-basic/blob/master/client/GameClient.js

Right now, I'm pretty sure these methods are the only way for a typescript user to tell when entities are added and removed. Without these,
I can't assign Identities to the entities in the game.